+++ test/pages/journal/entry_detail_page_test.dart
@@ -98,8 +98,7 @@ void main() {
 
       when(
         () => mockJournalDb.watchJournalEntities(
-          types:
-              defaultTypes.map((e) => e.typeName).whereType<String>().toList(),
+          types: entryTypes.map((e) => e.typeName).whereType<String>().toList(),
           starredStatuses: [true, false],
           privateStatuses: [true, false],
           flaggedStatuses: [1, 0],